+++ rust/lsp-lib/examples/json-lsp-plugin/src/main.rs
@@ -2,7 +2,6 @@ extern crate xi_lsp_lib;
 use xi_lsp_lib::{start_mainloop, LSPPlugin};
 
 fn main() {
-    
     // TODO: Make path to plugin configurable
     let mut plugin = LSPPlugin::new(
         "/usr/local/bin/vscode-json-languageserver",
@@ -10,7 +9,7 @@ fn main() {
         vec!["json".to_string(), "jsonc".to_string()],
         None,
         "json",
-    ); 
-    
+    );
+
     start_mainloop(&mut plugin);
 }
